requirements.txt
django==5.2.4
djangorestframework==3.15.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
whitenoise==6.9.0
django-cors-headers==4.7.0
dj-database-url==3.0.1
psutil==5.9.8
prometheus-client==0.20.0
requests==2.32.4
channels==4.1.0
channels-redis==4.2.0
drf-spectacular==0.27.2
uvicorn==0.32.1
daphne==4.2.1
